+++ app/src/main/java/com/sholatapp/ui/screens/TilawahScreen.kt
@@ -372,4 +372,3 @@ private fun MiniStatCard(
         }
     }
 }
-}